trend reference only if reference selected
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@121348 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 4744e1932c1d04d8bb82fbc6acd11b3551934d9e

+++ Online/Online/Presenter/src/CreateTrendingHistogramDialog.cpp
@@ -95,7 +95,7 @@ void CreateTrendingHistogramDialog::ok() {
 
       bool result = m_histdb -> commit() ;
 
-      sprintf( message, "Committed %d tags for trending file %s, status %d.", tags.size(), fileName.c_str(), result );
+      sprintf( message, "Committed %d tags for trending file %s, status %d.", (int)tags.size(), fileName.c_str(), result );
       *m_output = std::string( message );
       std::cout << message << std::endl;
     }
